--- src/main/java/com/kgm/PreOrder/domain/Follow.java
@@ -1,31 +1,0 @@
-package com.kgm.PreOrder.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "follow")
-public class Follow {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "follow_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "follower_id")
-    private User follower;
-
-    @ManyToOne
-    @JoinColumn(name = "following_id")
-    private User following;
-
-    public Follow(User follower, User following){
-        this.follower = follower;
-        this.following = following;
-    }
-}
