--- src/main/java/com/kgm/PreOrder/domain/Reply.java
@@ -1,32 +1,0 @@
-package com.kgm.PreOrder.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "reply")
-public class Reply {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "reply_id")
-    private Long id;
-
-    private String comment;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    private LocalDateTime date;
-
-}
