--- src/main/java/com/kgm/PreOrder/domain/Reply_like.java
@@ -1,29 +1,0 @@
-package com.kgm.PreOrder.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "reply_like")
-public class Reply_like {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "reply_like_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "reply_id")
-    private Reply reply;
-
-    private LocalDateTime date;
-}
