+++ src/main/java/com/kgm/preorder/PreOrderApplication.java
@@ -1,4 +1,4 @@
-package com.kgm.PreOrder;
+package com.kgm.preorder;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
