+++ src/main/java/com/kgm/preorder/controller/PostController.java
@@ -1,0 +1,9 @@
+package com.kgm.preorder.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class PostController {
+}
