+++ src/main/java/com/kgm/preorder/controller/UserController.java
@@ -1,0 +1,14 @@
+package com.kgm.preorder.controller;
+
+import com.kgm.preorder.domain.User;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class UserController {
+
+
+}
