+++ src/main/java/com/kgm/preorder/domain/Post.java
@@ -1,0 +1,14 @@
+package com.kgm.preorder.domain;
+
+import java.time.LocalDateTime;
+
+public class Post {
+
+    private Long id;
+
+    private String content;
+
+    private User user;
+
+    private LocalDateTime date;
+}
