+++ src/main/java/com/kgm/preorder/domain/Post_like.java
@@ -1,0 +1,15 @@
+package com.kgm.preorder.domain;
+
+import java.time.LocalDateTime;
+
+
+public class Post_like {
+
+    private Long id;
+
+    private User user;
+
+    private Post post;
+
+    private LocalDateTime date;
+}
