+++ src/main/java/com/kgm/preorder/domain/Reply.java
@@ -1,0 +1,17 @@
+package com.kgm.preorder.domain;
+
+import java.time.LocalDateTime;
+
+public class Reply {
+
+    private Long id;
+
+    private String comment;
+
+    private User user;
+
+    private Post post;
+
+    private LocalDateTime date;
+
+}
