+++ src/main/java/com/kgm/preorder/domain/Reply_like.java
@@ -1,0 +1,14 @@
+package com.kgm.preorder.domain;
+
+import java.time.LocalDateTime;
+
+public class Reply_like {
+
+    private Long id;
+
+    private User user;
+
+    private Reply reply;
+
+    private LocalDateTime date;
+}
