+++ src/main/java/com/kgm/preorder/domain/User.java
@@ -1,19 +1,9 @@
-package com.kgm.PreOrder.domain;
+package com.kgm.preorder.domain;
 
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
 import java.time.LocalDateTime;
 
-@Entity
-@Getter @Setter
-@Table(name = "user")
 public class User {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
     private Long id;
 
     private String email;
