+++ src/main/java/com/kgm/preorder/repository/UserMapper.java
@@ -1,0 +1,7 @@
+package com.kgm.preorder.repository;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface UserMapper {
+}
