+++ src/main/java/com/kgm/preorder/service/PostService.java
@@ -1,0 +1,4 @@
+package com.kgm.preorder.service;
+
+public interface PostService {
+}
