+++ src/main/java/com/kgm/preorder/serviceImpl/UserServiceImpl.java
@@ -1,0 +1,9 @@
+package com.kgm.preorder.serviceImpl;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl {
+}
