+++ src/main/resources/application.properties
@@ -1,17 +1,19 @@
-# ?????? ?? ??
 spring.datasource.url=jdbc:mysql://localhost:3306/preorder?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC
 spring.datasource.username=root
 spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
-# Hibernate ??
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-spring.jpa.properties.hibernate.default-storage-engine=innodb
+mybatis.type-aliases-package=com.kgm.preorder.domain
+mybatis.configuration.map-underscore-to-camel-case=true
 
-
-# JPA Auditing ???
-spring.data.jpa.repositories.enabled=true
-spring.jpa.properties.hibernate.jdbc.time_zone=UTC
+# email
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=<gmail kgm7642@gmail.com>
+spring.mail.password=<gmail kkggmm15987>
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
+spring.mail.properties.mail.smtp.starttls.required=true
+spring.mail.properties.mail.smtp.connectiontimeout=5000
+spring.mail.properties.mail.smtp.timeout=5000
+spring.mail.properties.mail.smtp.writetimeout=5000
